Cleaned up index page. Added meetups page.
